+++ analysis/filter/MergeFilter/src/main/kotlin/de/maibornwolff/codecharta/filter/mergefilter/MergeFilter.kt
@@ -34,6 +34,7 @@ import de.maibornwolff.codecharta.serialization.ProjectSerializer
 import picocli.CommandLine
 import java.io.File
 import java.util.concurrent.Callable
+import mu.KotlinLogging
 
 @CommandLine.Command(name = "merge",
         description = ["merges multiple cc.json files"],
@@ -42,7 +43,7 @@ class MergeFilter : Callable<Void?> {
     @CommandLine.Option(names = ["-h", "--help"], usageHelp = true, description = ["displays this help and exits"])
     var help: Boolean = false
 
-    @CommandLine.Parameters(arity = "1..*", paramLabel = "FILE", description = ["files to merge"])
+    @CommandLine.Parameters(arity = "1..*", paramLabel = "FILE or FOLDER", description = ["files to merge"])
     private var sources: Array<File> = arrayOf()
 
     @CommandLine.Option(names = ["-a", "--add-missing"], description = ["enable adding missing nodes to reference"])
@@ -60,6 +61,8 @@ class MergeFilter : Callable<Void?> {
     @CommandLine.Option(names = ["--ignore-case"], description = ["ignores case when checking node names"])
     private var ignoreCase = false
 
+    private val logger = KotlinLogging.logger {}
+
     override fun call(): Void? {
         val nodeMergerStrategy =
                 when {
@@ -68,9 +71,21 @@ class MergeFilter : Callable<Void?> {
                     else -> throw IllegalArgumentException("Only one merging strategy must be set")
                 }
 
-        val srcProjects = sources
-                .map { it.bufferedReader() }
-                .map { ProjectDeserializer.deserializeProject(it) }
+        val sourceFiles = mutableListOf<File>()
+        for(source in sources){
+            sourceFiles.addAll(getFilesInFolder(source))
+        }
+
+        val srcProjects = sourceFiles
+                .mapNotNull {
+                    val bufferedReader = it.bufferedReader()
+                    try {
+                        ProjectDeserializer.deserializeProject(bufferedReader)
+                    } catch (e: Exception){
+                        logger.warn("${it.name} is not a valid project file and is therefore skipped.")
+                        null
+                    }
+                }
 
         val mergedProject = ProjectMerger(srcProjects, nodeMergerStrategy).merge()
 
@@ -78,6 +93,11 @@ class MergeFilter : Callable<Void?> {
         ProjectSerializer.serializeProject(mergedProject, writer)
 
         return null
+    }
+
+    private fun getFilesInFolder(folder: File): List<File>{
+        val files =  folder.walk().filter{ !it.name.startsWith(".") && !it.isDirectory}
+        return files.toList()
     }
 
     companion object {